Added SeekBar to make percentage more user friendly
SeekBar on top of the last fragment enables user to define the
percentage *on fly*. It is also a good example of static and functional
usage of the setPercentage.
Snackbar needed a try-catch as it is throwing an
```
java.lang.NullPointerException: Attempt to invoke virtual method 'void
android.widget.TextView.setText(java.lang.CharSequence)' on a null
object reference
```
while on InstantRun and I could not be bothered to fix the real reason.