Fixed Coverity error UNINTENDED_INTEGER_DIVISION
Coverity Scan found an error:
```
CID 105392 (#1 of 1): Result is not floating-point
(UNINTENDED_INTEGER_DIVISION)integer_division: Dividing integer
expressions width and 2 * squareSize, and then converting the integer
quotient to type double. Any remainder, or fractional part of the
quotient, is ignored.
```

The solution is simply to cast one of the arguments to double. This
commit does that.